+++ src/model.cpp
@@ -1,7 +1,7 @@
 #include "include/model.hpp"
 
 cv::Mat randomNormal(int rows, int cols) {
-    cv::Mat result(rows, cols, CV_32FC1);
+    cv::Mat result(rows, cols, CV_64FC1);
     cv::RNG rng(static_cast<unsigned int>(std::time(0)));
 
     for (int i = 0; i < rows; i++) {
@@ -20,10 +20,10 @@ mazeNet::mazeNet(int in, int hide, int out) {
     this->output_size = out;
 
     w1 = randomNormal(input_size, hidden_size);
-    b1 = cv::Mat::zeros(1, hidden_size, CV_32FC1);
+    b1 = cv::Mat::zeros(1, hidden_size, CV_64FC1);
 
     w2 = randomNormal(hidden_size, output_size);
-    b2 = cv::Mat::zeros(1, output_size, CV_32FC1);
+    b2 = cv::Mat::zeros(1, output_size, CV_64FC1);
 }
 
 
@@ -32,9 +32,10 @@ cv::Mat mazeNet :: forward (const cv::Mat & X) {
     z1 = utils.sum(z1, b1);
 
     a1 = utils.relu(z1);
-
+    
     z2 = utils.dot(a1, w2);
     z2 = utils.sum(z2, b2);
+    // cout<<result<<endl;
     result = utils.softmax(z2);
 
     return result;
@@ -43,7 +44,7 @@ cv::Mat mazeNet :: forward (const cv::Mat & X) {
 
 void mazeNet :: backward (const cv::Mat & X_train ,const cv::Mat & y_train, const cv::Mat & y_pred, float learning_rate) {
 
-    cv::Mat y_pred_one_hot(y_pred.rows, y_pred.cols, CV_32FC1, cv::Scalar(0.0));  // Initialize with zeros
+    cv::Mat y_pred_one_hot(y_pred.rows, y_pred.cols, CV_64FC1, cv::Scalar(0.0));  // Initialize with zeros
 
     for (int i = 0; i < y_pred.rows; i++) {
         // Find the index of the maximum value in the current row (argmax)
@@ -58,7 +59,7 @@ void mazeNet :: backward (const cv::Mat & X_train ,const cv::Mat & y_train, cons
     // Calculate gradients for the output layer
     cv::Mat a1_Transpose = a1.t();
     cv::Mat w2_Gradient = utils.dot(a1_Transpose ,lossGradient);
-    cv::Mat b2_Gradient = utils.dot(cv::Mat::ones(1, lossGradient.rows, CV_32FC1) ,lossGradient);
+    cv::Mat b2_Gradient = utils.dot(cv::Mat::ones(1, lossGradient.rows, CV_64FC1) ,lossGradient);
 
     // Calculate gradients for the hidden layer
     cv::Mat z1_Gradient = utils.dot(lossGradient ,w2.t());
@@ -67,7 +68,7 @@ void mazeNet :: backward (const cv::Mat & X_train ,const cv::Mat & y_train, cons
 
     cv::Mat X_Transpose = X_train.t();
     cv::Mat w1_Gradient = utils.dot(X_Transpose ,z1_Gradient);
-    cv::Mat b1_Gradient = utils.dot(cv::Mat::ones(1, z1_Gradient.rows, CV_32FC1) ,z1_Gradient);
+    cv::Mat b1_Gradient = utils.dot(cv::Mat::ones(1, z1_Gradient.rows, CV_64FC1) ,z1_Gradient);
     // Update weights and biases using gradients and learning rate
     w1 -= learning_rate * w1_Gradient;
     b1 -= learning_rate * b1_Gradient;
